+++ pico-rv32ima/CMakeLists.txt
@@ -1,3 +1,4 @@
+add_subdirectory(petitfatfs)
 add_subdirectory(vga)
 
 add_executable(pico-rv32ima
@@ -9,8 +10,6 @@ add_executable(pico-rv32ima
 	emulator.c
 	get_serial.c
 	usb_descriptors.c
-	pff.c
-	mmcbbp.c
 )
 
 # Add the standard library and FatFS/SPI to the build
@@ -24,7 +23,9 @@ target_link_libraries(pico-rv32ima
 	hardware_clocks
 	tinyusb_device 
 	tinyusb_board
+	petitfatfs
 	vga
+
 )
 
 target_include_directories(pico-rv32ima PUBLIC ${CMAKE_CURRENT_LIST_DIR}/tusb_inc )
