+++ pico-rv32ima/petitfatfs/CMakeLists.txt
@@ -1,0 +1,10 @@
+add_library(petitfatfs
+	pff.c
+    mmcbbp.c
+)
+
+target_include_directories(petitfatfs PUBLIC
+	.
+)
+
+target_link_libraries(petitfatfs pico_stdlib hardware_spi)
